Deduplicating some code and resolving TODOs
