+++ src/layout/FileUpload/AttachmentSummaryComponent.tsx
@@ -4,7 +4,7 @@ import { Grid, makeStyles, Typography } from '@material-ui/core';
 
 import { useAppSelector } from 'src/common/hooks/useAppSelector';
 import { getLanguageFromKey } from 'src/language/sharedLanguage';
-import type { IAttachment } from 'src/shared/resources/attachments';
+import { useUploaderSummaryData } from 'src/layout/FileUpload/shared/summary';
 import type { LayoutNodeFromType } from 'src/utils/layout/hierarchy.types';
 
 export interface IAttachmentSummaryComponent {
@@ -19,20 +19,16 @@ const useStyles = makeStyles({
 });
 
 export function AttachmentSummaryComponent({ targetNode }: IAttachmentSummaryComponent) {
-  // PRIORITY: Find out if summary data from the component class is needed here at all?
   const classes = useStyles();
-  const attachments: IAttachment[] | undefined = useAppSelector(
-    (state) => state.attachments.attachments[targetNode.item.id],
-  );
+  const attachments = useUploaderSummaryData(targetNode);
   const language = useAppSelector((state) => state.language.language);
-  const isEmpty = !attachments || attachments.length < 1;
   return (
     <Grid
       item
       xs={12}
       data-testid={'attachment-summary-component'}
     >
-      {isEmpty ? (
+      {attachments.length === 0 ? (
         <Typography
           variant='body1'
           className={classes.emptyField}
@@ -41,16 +37,14 @@ export function AttachmentSummaryComponent({ targetNode }: IAttachmentSummaryCom
           {getLanguageFromKey('general.empty_summary', language || {})}
         </Typography>
       ) : (
-        attachments.map((attachment) => {
-          return (
-            <Typography
-              key={attachment.id}
-              variant='body1'
-            >
-              {attachment.name}
-            </Typography>
-          );
-        })
+        attachments.map((attachment) => (
+          <Typography
+            key={attachment.id}
+            variant='body1'
+          >
+            {attachment.name}
+          </Typography>
+        ))
       )}
     </Grid>
   );
