+++ src/layout/FileUpload/index.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
 
-import { useAppSelector } from 'src/common/hooks/useAppSelector';
 import { AttachmentSummaryComponent } from 'src/layout/FileUpload/AttachmentSummaryComponent';
 import { FileUploadComponent } from 'src/layout/FileUpload/FileUploadComponent';
-import {
-  attachmentNamesFromComponentId,
-  attachmentNamesFromUuids,
-  extractListFromBinding,
-} from 'src/layout/FileUpload/shared/summary';
+import { useUploaderSummaryData } from 'src/layout/FileUpload/shared/summary';
 import { FormComponent } from 'src/layout/LayoutComponent';
 import type { PropsFromGenericComponent } from 'src/layout';
 import type { SummaryRendererProps } from 'src/layout/LayoutComponent';
@@ -22,21 +17,10 @@ export class FileUpload extends FormComponent<'FileUpload'> {
     return false;
   }
 
-  private useSummaryData(node: LayoutNodeFromType<'FileUpload'>): string[] {
-    const formData = useAppSelector((state) => state.formData.formData);
-    const attachments = useAppSelector((state) => state.attachments.attachments);
-
-    const listBinding = node.item.dataModelBindings?.list;
-    if (listBinding) {
-      const values = extractListFromBinding(formData, listBinding);
-      return attachmentNamesFromUuids(node.item.id, values, attachments);
-    }
-
-    return attachmentNamesFromComponentId(node.item.id, attachments);
-  }
-
   useDisplayData(node: LayoutNodeFromType<'FileUpload'>): string {
-    return this.useSummaryData(node).join(', ');
+    return useUploaderSummaryData(node)
+      .map((a) => a.name)
+      .join(', ');
   }
 
   renderSummary({ targetNode }: SummaryRendererProps<'FileUpload'>): JSX.Element | null {
