+++ src/layout/FileUploadWithTag/index.tsx
@@ -1,11 +1,6 @@
 import React from 'react';
 
-import { useAppSelector } from 'src/common/hooks/useAppSelector';
-import {
-  attachmentNamesFromComponentId,
-  attachmentNamesFromUuids,
-  extractListFromBinding,
-} from 'src/layout/FileUpload/shared/summary';
+import { useUploaderSummaryData } from 'src/layout/FileUpload/shared/summary';
 import { AttachmentWithTagSummaryComponent } from 'src/layout/FileUploadWithTag/AttachmentWithTagSummaryComponent';
 import { FileUploadWithTagComponent } from 'src/layout/FileUploadWithTag/FileUploadWithTagComponent';
 import { FormComponent } from 'src/layout/LayoutComponent';
@@ -22,21 +17,10 @@ export class FileUploadWithTag extends FormComponent<'FileUploadWithTag'> {
     return false;
   }
 
-  private useSummaryData(node: LayoutNodeFromType<'FileUploadWithTag'>): string[] {
-    const formData = useAppSelector((state) => state.formData.formData);
-    const attachments = useAppSelector((state) => state.attachments.attachments);
-
-    const listBinding = node.item.dataModelBindings?.list;
-    if (listBinding) {
-      const values = extractListFromBinding(formData, listBinding);
-      return attachmentNamesFromUuids(node.item.id, values, attachments);
-    }
-
-    return attachmentNamesFromComponentId(node.item.id, attachments);
-  }
-
   useDisplayData(node: LayoutNodeFromType<'FileUploadWithTag'>): string {
-    return this.useSummaryData(node).join(', ');
+    return useUploaderSummaryData(node)
+      .map((a) => a.name)
+      .join(', ');
   }
 
   renderSummary({ targetNode }: SummaryRendererProps<'FileUploadWithTag'>): JSX.Element | null {
